backend/requirements.txt
flask
flask-cors
transformers
torch
pandas
numpy
scikit-learnrequests